Make price time filtering non-intrusive with feature flag.
