+++ VirtoCommerce.PricingModule.Data/VirtoCommerce.PricingModule.Data.csproj
@@ -109,7 +109,6 @@
     <Compile Include="Repositories\IPricingRepository.cs" />
     <Compile Include="Search\ProductPriceDocumentBuilder.cs" />
     <Compile Include="Search\ProductPriceDocumentChangesProvider.cs" />
-    <Compile Include="Services\PriceExtensions.cs" />
     <Compile Include="Services\PricingSearchServiceImpl.cs" />
     <Compile Include="Services\PricingServiceImpl.cs" />
   </ItemGroup>
