+++ VirtoCommerce.PricingModule.Web/Localizations/en.VirtoCommerce.Pricing.json
@@ -56,6 +56,8 @@
           "list-price": "List price",
           "sale-price": "Sale price",
           "min-quantity": "Min. quantity",
+          "startDate": "Valid from",
+          "endDate": "Valid to",
           "last-modified": "Modified"
         }
       },
