+++ VirtoCommerce.PricingModule.Web/Scripts/blades/item/item-prices.tpl.html
@@ -31,8 +31,8 @@
                         { name: 'list', displayName: 'pricing.blades.item-prices.labels.list-price', editableCellTemplate: 'list-cellTextEditor', validators: {listValidator: true, required: true}, cellTemplate: 'priceCellTitleValidator', enableCellEdit: true, enableHiding: false },
                         { name: 'sale', displayName: 'pricing.blades.item-prices.labels.sale-price', editableCellTemplate: 'sale-cellTextEditor', validators: {saleValidator: true}, cellTemplate: 'priceCellTitleValidator', enableCellEdit: true },
                         { name: 'minQuantity', displayName: 'pricing.blades.item-prices.labels.min-quantity', editableCellTemplate: 'minQuantity-cellTextEditor', validators: {minQuantityForPricelistValidator: true, required: true}, cellTemplate: 'ui-grid/cellTitleValidator', enableCellEdit: true, enableHiding: false },
-                        { name: 'startDate', displayName: 'pricing.blades.item-prices.labels.startDate', cellTooltip: true, cellTemplate: 'ui-grid/dateCellTemplate',editableCellTemplate: 'date-cellTextEditor', enableCellEdit: true },
-                        { name: 'endDate', displayName: 'pricing.blades.item-prices.labels.endDate', cellTooltip: true, cellTemplate: 'ui-grid/dateCellTemplate', editableCellTemplate: 'date-cellTextEditor', enableCellEdit: true }                        
+                        { name: 'startDate', displayName: 'pricing.blades.item-prices.labels.startDate', cellTooltip: true, cellTemplate: 'ui-grid/dateCellTemplate', editableCellTemplate: 'date-cellTextEditor', enableCellEdit: true, visible: false },
+                        { name: 'endDate', displayName: 'pricing.blades.item-prices.labels.endDate', cellTooltip: true, cellTemplate: 'ui-grid/dateCellTemplate', editableCellTemplate: 'date-cellTextEditor', enableCellEdit: true, visible: false }
                     ]})">
                 <div class="table-wrapper" ng-if="blade.currentEntities.length" ng-init="setForm(formScope);">
                     <div ui-grid="gridOptions" ui-grid-cellNav ui-grid-edit ui-grid-validate ui-grid-auto-resize ui-grid-save-state ui-grid-selection ui-grid-resize-columns ui-grid-move-columns ui-grid-height></div>
